YOLOv8_TensorRT/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yolov8-trt"
version = "0.1.0"
description = "TensorRT-accelerated YOLOv8 COCO object detection"
requires-python = ">=3.8"
dependencies = [
    "numpy",
    "opencv-python",
    "torch",
    "ultralytics",
    "onnx",
    "onnxsim",
    "onnxruntime-gpu",
    "tensorrt",
    "pycuda",
]

[tool.setuptools.packages.find]
where = ["src"]
